@@ -1,11 +1,11 @@
 name: dotted_line
 description: This package allows you to draw dotted lines with Flutter. You can draw a beautiful dotted line.
-version: 2.0.2
+version: 3.0.0
 homepage: https://github.com/umechanhika/dotted_line
 repository: https://github.com/umechanhika/dotted_line
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
